+++ app/interfaces/ajax.php
@@ -742,31 +742,6 @@ if ($accion == "consistencia_cliente_asunto") {
 		$password = $_POST['password'];
 	}
 	echo PasswordStrength::Rate($password);
-} else if ($accion == 'permite_asunto_independiente') {
-	$response = array('error' => false, 'error_glosa' => '');
-	if ($codigo_cliente != '') {
-		$Cliente = new Cliente($sesion);
-		$loaded = Conf::GetConf($sesion, 'CodigoSecundario') ? $Cliente->LoadByCodigoSecundario($codigo_cliente) : $Cliente->LoadByCodigo($codigo_cliente);
-		if ($loaded) {
-			$Asunto = new Asunto($sesion);
-			if (!empty($id_asunto)) {
-				if (!$Asunto->Load($id_asunto)) {
-					$response['error_glosa'] = utf8_encode('Código asunto inválido');
-				}
-			}
-			if (empty($response['error_glosa'])) {
-				$response['permite_asunto_independiente'] = $Asunto->permiteAsuntoIndependiente($Cliente->fields['codigo_cliente']);
-			}
-		} else {
-			$response['error_glosa'] = utf8_encode('Código cliente inválido');
-		}
-	} else {
-		$response['error_glosa'] = utf8_encode('Código cliente vacío');
-	}
-	if (!empty($response['error_glosa'])) {
-		$response['error'] = true;
-	}
-	echo json_encode($response);
 } else {
 	echo "ERROR AJAX. Acción: $accion";
 }